analyzer-server/requirements.txt
fastapi[standard]
uvicorn
python-multipart
openai
python-dotenv
